requirements.txt
-e .
appdirs==1.4.3
certifi==2018.1.18
chardet==3.0.4
Django==1.11.7
pytest-parallel
pytest-django
pytest-pythonpath
django-widget-tweaks==1.4.1
idna==2.6
packaging==16.8
pyparsing==2.2.0
pyshorteners==1.0.0
pytz==2017.3
requests==2.18.4
six==1.10.0
urllib3==1.22
djangorestframework==3.7.7
